implement problem options
