+++ able/lib/main.dart
@@ -1,8 +1,9 @@
 import 'package:able/main.dart';
 import 'package:flutter/material.dart';
-import './style.dart' as style;
-import 'package:http/http.dart' as http;
-import 'dart:convert';
+import './styles.dart' as style;
+import './variables.dart' as variable;
+import './catalogPage.dart' as catalog;
+import './testPage.dart' as test;
 import 'package:provider/provider.dart';
 
 void main() {
@@ -20,115 +21,8 @@ void main() {
   );
 }
 
-class StoreAboutData extends ChangeNotifier {
-  var name = '비에이블투';
-  var data = [];
-  var ch = 1;
-  var page = 0;
-
-  nameHome() {
-    name = '비에이블투';
-    notifyListeners();
-  }
-
-  nameLogic() {
-    name = '논리회로';
-    notifyListeners();
-  }
-
-  nameElectric() {
-    name = '전자회로';
-    notifyListeners();
-  }
-
-  nameRequest() {
-    name = '질문게시판';
-    notifyListeners();
-  }
-
-  getData() async {
-    var chData;
-    var chData_;
-    if (ch == 1) {
-      name == '논리회로'
-        ? chData = await http.get( Uri.parse('https://sphcsphc.github.io/logic_circuit/ch01/ch01.json') )
-        : chData = await http.get( Uri.parse('https://sphcsphc.github.io/logic_circuit/ch01/ch01.json') );
-    }
-    if (ch == 2) {
-      name == '논리회로'
-        ? chData = await http.get( Uri.parse('https://sphcsphc.github.io/logic_circuit/ch02/ch02.json') )
-        : chData = await http.get( Uri.parse('https://sphcsphc.github.io/logic_circuit/ch01/ch01.json') );
-    }
-    if (ch == 3) {
-      name == '논리회로'
-        ? chData = await http.get( Uri.parse('https://sphcsphc.github.io/logic_circuit/ch03/ch03.json') )
-        : chData = await http.get( Uri.parse('https://sphcsphc.github.io/logic_circuit/ch01/ch01.json') );
-    }
-    if (ch == 4) {
-      name == '논리회로'
-        ? chData = await http.get( Uri.parse('https://sphcsphc.github.io/logic_circuit/ch04/ch04.json') )
-        : chData = await http.get( Uri.parse('https://sphcsphc.github.io/logic_circuit/ch01/ch01.json') );
-    }
-    chData_ = jsonDecode(chData.body);
-    data = chData_;
-    page = 0;
-    notifyListeners();
-  }
-
-  chOne() {
-    ch = 1;
-    getData();
-  }
-  chTwo() {
-    ch = 2;
-    getData();
-  }
-  chThree() {
-    ch = 3;
-    getData();
-  }
-  chFour() {
-    ch = 4;
-    getData();
-  }
-
-  previousPage() {
-    page--;
-    notifyListeners();
-  }
-  nextPage() {
-    page++;
-    notifyListeners();
-  }
-}
-
-class StoreAboutTab extends ChangeNotifier {
-  var tab = 0;
-
-  goHome() {
-    tab = 0;
-    notifyListeners();
-  }
-
-  goTest() {
-    tab = 1;
-    notifyListeners();
-  }
-
-  goRequest() {
-    tab = 2;
-    notifyListeners();
-  }
-}
-
-class MyApp extends StatefulWidget {
+class MyApp extends StatelessWidget {
   const MyApp({Key? key}) : super(key: key);
-
-  @override
-  State<MyApp> createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) {
@@ -167,153 +61,6 @@ class _MyAppState extends State<MyApp> {
   }
 }
 
-class Catalog extends StatelessWidget {
-  const Catalog({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('바로가기'),
-        actions: [
-          IconButton(
-              icon: Icon(Icons.arrow_back_ios_new),
-              onPressed: (){
-                Navigator.pop(context);
-              }
-          )
-        ]
-      ),
-      body: ListView(
-        children: const [
-          CatalogBody(),
-        ]
-      )
-    );
-  }
-}
-
-class CatalogBody extends StatelessWidget {
-  const CatalogBody({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children : [
-        Container(
-          width : 350, height : 150,
-          margin: EdgeInsets.all(30),
-          padding: EdgeInsets.fromLTRB(20, 20, 30, 40),
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(50),
-            border: Border.all(color: Colors.black12, width: 2)
-          ),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Row(
-                children: [
-                  Icon(Icons.power),
-                  TextButton(onPressed: (){
-                    context.read<StoreAboutData>().nameLogic();
-                    context.read<StoreAboutTab>().goTest();
-                    showDialog(
-                        context: context,
-                        builder: (context){ return SetChapter(); }
-                    );
-                  }, child: Text('논리회로')),
-                ],
-              ),
-              Row(
-                children: [
-                  Icon(Icons.cable),
-                  TextButton(onPressed: (){
-                    context.read<StoreAboutData>().nameElectric();
-                    context.read<StoreAboutTab>().goTest();
-                    showDialog(
-                        context: context,
-                        builder: (context){ return SetChapter(); }
-                    );
-                  }, child: Text('전자회로')),
-                ],
-              ),
-            ],
-          )
-        ),
-        Container(
-          width : 350, height : 150,
-          margin: EdgeInsets.all(30),
-          padding: EdgeInsets.fromLTRB(20, 20, 30, 40),
-          decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(50),
-              border: Border.all(color: Colors.black12, width: 2)
-          ),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Row(
-                children: [
-                  Icon(Icons.question_answer),
-                  TextButton(onPressed: (){
-                    context.read<StoreAboutData>().nameRequest();
-                    context.read<StoreAboutTab>().goRequest();
-                    Navigator.pop(context);
-                  }, child: Text('질문게시판')),
-                ],
-              )
-            ],
-          ),
-        ),
-      ]
-    );
-  }
-}
-
-class SetChapter extends StatelessWidget {
-  const SetChapter({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      content: Text('몇장 볼꺼야?'),
-      actions: [
-        Center(
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: [
-              ElevatedButton(onPressed: (){
-                context.read<StoreAboutData>().chOne();
-                Navigator.push(context, MaterialPageRoute(
-                    builder: (BuildContext context) =>
-                        MyApp()));
-              }, child: Text('1')),
-              ElevatedButton(onPressed: (){
-                context.read<StoreAboutData>().chTwo();
-                Navigator.push(context, MaterialPageRoute(
-                    builder: (BuildContext context) =>
-                        MyApp()));
-              }, child: Text('2')),
-              ElevatedButton(onPressed: (){
-                context.read<StoreAboutData>().chThree();
-                Navigator.push(context, MaterialPageRoute(
-                    builder: (BuildContext context) =>
-                        MyApp()));
-              }, child: Text('3')),
-              ElevatedButton(onPressed: (){
-                context.read<StoreAboutData>().chFour();
-                Navigator.push(context, MaterialPageRoute(
-                    builder: (BuildContext context) =>
-                        MyApp()));
-              }, child: Text('4')),
-            ],
-          ),
-        )
-      ],
-    );
-  }
-}
-
-
 class Home extends StatelessWidget {
   const Home({Key? key}) : super(key: key);
 
@@ -323,96 +70,6 @@ class Home extends StatelessWidget {
   }
 }
 
-
-class Test extends StatefulWidget {
-  const Test({Key? key}) : super(key: key);
-
-  @override
-  State<Test> createState() => _TestState();
-}
-
-class _TestState extends State<Test> {
-
-  @override
-  Widget build(BuildContext context) {
-    if (context.watch<StoreAboutData>().data.isNotEmpty){
-      return Container(
-        margin: EdgeInsets.all(30),
-        child: ShowProblem()
-      );
-    } else {
-      return Center( child: Text('로딩중') );
-    }
-  }
-}
-
-class ShowProblem extends StatelessWidget {
-  const ShowProblem({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: [
-        context.watch<StoreAboutData>().data[context.watch<StoreAboutData>().page]['problem'].runtimeType == String
-          ? Image.network(context.watch<StoreAboutData>().data[context.watch<StoreAboutData>().page]['problem'])
-          : Image.file(context.watch<StoreAboutData>().data[context.watch<StoreAboutData>().page]['problem']),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: [
-            context.watch<StoreAboutData>().page == 0
-              ? Text('처음')
-              : ElevatedButton(onPressed: (){
-                context.read<StoreAboutData>().previousPage();
-                }, child: Text('이전')),
-            ElevatedButton(onPressed: (){
-              showDialog(
-                context: context,
-                builder: (context){ 
-                  return ListView(
-                    children: const [
-                      ShowSolve(); 
-                    ]
-                  );
-                }
-              );
-            }, child: Text('정답보기')),
-            context.watch<StoreAboutData>().page == context.watch<StoreAboutData>().data.length - 1
-              ? Text('마지막')
-              : ElevatedButton(onPressed: (){
-                context.read<StoreAboutData>().nextPage();
-                }, child: Text('다음')),
-          ],
-        )
-      ]
-    );
-  }
-}
-
-class ShowSolve extends StatelessWidget {
-  const ShowSolve({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        context.watch<StoreAboutData>().data[context.watch<StoreAboutData>().page]['solution'].runtimeType == String
-          ? Image.network(context.watch<StoreAboutData>().data[context.watch<StoreAboutData>().page]['solution'])
-          : Image.file(context.watch<StoreAboutData>().data[context.watch<StoreAboutData>().page]['solution']),
-        RawMaterialButton(
-          onPressed: (){ Navigator.pop(context); },
-          elevation: 2.0,
-          fillColor: Colors.white,
-          child: Icon(Icons.close, size: 15.0,),
-          padding: EdgeInsets.all(15.0),
-          shape: CircleBorder(),
-        ),
-      ],
-    );
-  }
-}
-
-
 class Request extends StatelessWidget {
   const Request({Key? key}) : super(key: key);
 
@@ -421,4 +78,3 @@ class Request extends StatelessWidget {
     return Container();
   }
 }
-
